refactor: Add i18n support for UI navigation
This commit adds support for internationalization (i18n) in the UI navigation component. The `ui.ts` file has been modified to include translations for the navigation items in English and Spanish. The `Layout.astro` and `Nav.astro` files have been updated to dynamically set the language attribute of the HTML element based on the current language. The `getLangFromUrl` and `useTranslations` functions have been added to the `utils.ts` file to handle language detection and translation respectively.

Ref: #123